requirements.txt
python-telegram-bot>=20.0
qrcode>=7.4
pillow>=10.0
